day_1/requirements.txt
affine==2.3.0
appnope==0.1.2
attrs==21.2.0
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
click-plugins==1.1.1
cligj==0.7.2
contextily==1.2.0
country-bounding-boxes==0.2.3
cycler==0.10.0
debugpy==1.5.0
decorator==5.1.0
entrypoints==0.3
Fiona==1.8.20
geographiclib==1.52
geopandas==0.10.1
geopy==2.2.0
grass-session==0.5
idna==3.3
imageio==2.10.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
iso3166==2.0.2
jedi==0.18.0
joblib==1.1.0
jupyter-client==7.0.6
jupyter-core==4.8.1
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
matplotlib-inline==0.1.3
mercantile==1.2.1
munch==2.5.0
nest-asyncio==1.5.1
numpy==1.21.2
osgb==1.0.0
overpy==0.6
pandas==1.3.3
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pygdal==3.3.2.10
Pygments==2.10.0
pyparsing==2.4.7
pyproj==3.2.1
python-dateutil==2.8.2
pytz==2021.3
pyzmq==22.3.0
rasterio==1.2.10
requests==2.26.0
Shapely==1.7.1
six==1.16.0
snuggs==1.4.7
tornado==6.1
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
xyzservices==2021.10.0
